+++ src/components/shareModule/ShareModule.ts
@@ -1,0 +1,7 @@
+export const copyLink = () => {
+  try {
+    alert('클립보드에 복사되었습니다.');
+  } catch (error) {
+    alert('클립보드 복사에 실패하였습니다.');
+  }
+};
