+++ src/pages/invite.tsx
@@ -1,4 +1,4 @@
-import InviteModal from '@src/components/inviteModal/InviteModal';
+import InviteModal from '@src/components/shareModule/InviteModal';
 
 export default function Invite() {
   return <InviteModal></InviteModal>;
