+++ src/pages/team/result.tsx
@@ -1,5 +1,5 @@
 import LogoTop from '@src/components/common/LogoTop';
-import TeamModal from '@src/components/teamModal/TeamModal';
+import TeamModal from '@src/components/shareModule/TeamModal';
 
 export default function Result() {
   return (
